+++ packages/build-mpa-config/src/generate/BaseGenerator.ts
@@ -61,7 +61,7 @@ export default class BaseGenerator {
         router: Boolean(routesFilePath),
       },
       routesFilePath: routesFilePath && relative(this.entryFolder, routesFilePath),
-      errorBoundary: false,
+      isMPA: true,
     };
 
     applyMethod('addRenderFile', getTemplate('runApp.ts', framework), `${this.runAppPath}.ts`, renderData);