+++ packages/build-user-config/src/userConfig/minify.js
@@ -2,17 +2,39 @@ const {
   ESBuildMinifyPlugin
 } = require('@builder/pack/deps/esbuild-loader');
 const { SWCMinifyPlugin } = require('@builder/webpack-plugin-swc');
+const TerserPlugin = require('@builder/pack/deps/terser-webpack-plugin');
+
+let logged = false;
+
+const TERSER_DEFAULT_OPTIONS = {
+  parallel: true,
+  extractComments: false,
+  terserOptions: {
+    output: {
+      ascii_only: true,
+      comments: 'some',
+      beautify: false,
+    },
+    mangle: true,
+  },
+};
 
 module.exports = (config, minify, context, { log }) => {
-  const { command } = context;
-  if ( minify && command === 'build') {
+  const { command, taskName } = context;
+  const includedInMinifyTasks = minify === true && (minify.activeInDev || []).includes(taskName);
+  if ( minify && (command === 'build' || includedInMinifyTasks)) {
     const minifierConfig = minify.type ? minify : { type: typeof minify === 'boolean' ? 'terser' : minify };
     const { type, options } = minifierConfig;
     const availableMinifier = ['terser', 'esbuild', 'swc'];
     if (!availableMinifier.includes(type)) {
-      log.info(`invalid minify value ${type}, available minifier is 'terser | esbuild'`);
+      if (!logged) {
+        logged = true;
+        log.info(`invalid minify value ${type}, available minifier is 'terser | esbuild'`);
+      }
       return;
     }
+    // Enable minimize for development, in production its default value is true
+    config.optimization.minimize(true);
     if (['swc', 'esbuild'].includes(type) && config.optimization.minimizers.has('TerserPlugin')) {
       config.optimization.minimizers.delete('TerserPlugin');
     }
@@ -38,7 +60,12 @@ module.exports = (config, minify, context, { log }) => {
         sync: false,
       };
       config.optimization.minimizer('SWC').use(SWCMinifyPlugin, [jscCompileOptions]);
-    } else if (config.optimization.minimizers.get('TerserPlugin')) {
+    } else {
+      if (!config.optimization.minimizers.has('TerserPlugin')) {
+        config.optimization
+          .minimizer('TerserPlugin')
+          .use(TerserPlugin, [TERSER_DEFAULT_OPTIONS]);
+      }
       // fallback to terser as default minification
       config.optimization.minimizer('TerserPlugin').tap(([terserPluginOptions]) => {
         return [{