+++ packages/create-app-shared/src/miniapp/index.ts
@@ -6,7 +6,7 @@ export { default as createHistory, initHistory } from './history';
 export { default as initAppLifeCycles} from './initAppLifeCycles';
 export { setHistory, getHistory, history } from '../storage';
 export { pathRedirect } from '../utils';
-export { default as getSearchParams} from '../getSearchParams';
+export { default as getSearchParams} from './getSearchParams';
 export { default as collectAppLifeCycle} from './collectAppLifeCycle';
 export { default as RuntimeModule, RuntimePlugin } from '../runtimeModule';
 export { default as createBaseApp } from '../createBaseApp';